fix: moved sort to lib/assistant
